+++ cli/procedures/route-add.js
@@ -2,61 +2,17 @@ import fs from "fs"
 import Template from "../Template"
 
 var force = false;
-var namespace = null;
+var namespace = "default";
 
-/**
- * @description Makes template.
- */
-async function routeAdd(method, path, handler) {
-  if (!NOTATIONS.endpointPath.test(path)) {
-    error(`A route path should start with a '/'. For example: "/posts".`);
-    return;
-  }
-  if (!NOTATIONS.controllerHandler.test(handler)) {
-    error(`A route handler should look like this: 'PostController.index'.`);
-    return;
-  }
-  const dir = `${process.cwd()}/app/routes`;
-  const file = `${namespace || "default"}.json`;
-  if (!fs.existsSync(`${process.cwd()}/app`)) {
-    fs.mkdirSync(`${process.cwd()}/app`);
-  }
-  if (!fs.existsSync(dir)) {
-    fs.mkdirSync(dir);
-  }
-  const endpoint = `${method.toUpperCase()} ${path}`;
-  if (fs.existsSync(`${dir}/${file}`)) {
-    const content = JSON.parse(fs.readFileSync(`${dir}/${file}`));
-    if (!Array.isArray(content.endpoints) && typeof content.endpoints === "object") {
-      if (content.endpoints.hasOwnProperty(endpoint)) {
-        error(`route ${endpoint} already exists!`);
-        return;
-      }
-      content.endpoints[endpoint] = [handler];
-    }
-    fs.writeFileSync(`${dir}/${file}`, JSON.stringify(content, null, 2));
-    return;
-  }
-  const context = {
-    endpoint: endpoint,
-    handler: handler
-  };
-  const template = new Template("routes-config");
-  try {
-    await template.render(context, `app/routes/${file}`, force);
-    console.log(`\x1b[32mDone.\x1b[0m`);
-  } catch (error) {
-    error(`(error) ${error}`);
-  }
-}
-
-
-module.exports.execute = routeAdd;
-module.exports.enabled = true;
-module.exports.requireEnvironment = false;
-module.exports.scope = "PROJECT";
-module.exports.command = "route-add <method> <path> <handler>";
-module.exports.options = {
+export const enabled = true;
+export const scope = "PROJECT";
+export const description = "adds a route configuration to the default namespace";
+export const parameters = {
+  method: "http method. For example: 'GET' or 'POST'",
+  path: "endpoint path starting with a slash. For example: '/posts'",
+  handler: "Controller handler. For example: 'PostController.index'"
+};
+export const options = {
   "-f, --force": {
     description: "forcefully perform action",
     execute: () => {
@@ -70,9 +26,49 @@ module.exports.options = {
     }
   }
 };
-module.exports.description = "Creates a new route.";
-module.exports.parameters = {
-  method: "http method. For example: 'GET' or 'POST'",
-  path: "endpoint path starting with a slash. For example: '/posts'",
-  handler: "Controller handler. For example: 'PostController.index'"
-};
+export default async (method, path, handler) => {
+  if (!NOTATIONS.endpointPath.test(path)) {
+    error(`Invalid path: '${path}'. A route path should look like this: "/posts/update".`);
+    return;
+  }
+  if (!NOTATIONS.controllerHandler.test(handler)) {
+    error(`A route handler should look like this: 'PostController.index'.`);
+    return;
+  }
+  const dir = project.getRoutesDir(true);
+  const file = `${namespace}.json`;
+  if (!fs.existsSync(project.getAppDir(true))) {
+    fs.mkdirSync(project.getAppDir(true));
+  }
+  if (!fs.existsSync(dir)) {
+    fs.mkdirSync(dir);
+  }
+  const endpoint = `${method.toUpperCase()} ${path}`;
+  const filepath = `${dir}/${file}`;
+  if (fs.existsSync(filepath)) {
+    const content = JSON.parse(fs.readFileSync(filepath));
+    if (!Array.isArray(content.endpoints) && typeof content.endpoints === "object") {
+      if (content.endpoints.hasOwnProperty(endpoint)) {
+        error(`route ${endpoint} already exists!`);
+        return;
+      }
+      content.endpoints[endpoint] = [handler];
+    }
+    fs.writeFile(filepath, JSON.stringify(content, null, 2), (err) => {
+      if (err !== null) return error(err);
+      print(`\x1b[32mSaved new route to '${namespace}' namespace: '${filepath}'.\x1b[0m`);
+    });
+    return;
+  }
+  const context = {
+    endpoint: endpoint,
+    handler: handler
+  };
+  const template = new Template("routes-config");
+  try {
+    await template.render(context, `app/routes/${file}`, force);
+    print(`\x1b[32mSaved new route to '${namespace}' namespace: '${filepath}'.\x1b[0m`);
+  } catch (error) {
+    error(`(error) ${error}`);
+  }
+}